backtesting/requirements.txt
fastapi
fastapi-utils
uvicorn[standard]
starlette
asyncio
pika
python-binance
binance_downloader
pandas
bta-lib
backtrader
backtrader[plotting]
matplotlib==3.2.2